requirements/base.txt
# Core Django
Django==6.0.2
django-htmx==1.27.0
django-unfold==0.80.2
django-anymail[resend]==14.0
django-storages[s3]==1.14.6
python-dotenv==1.2.1

# Database
dj-database-url==3.1.2
psycopg2-binary==2.9.11

# File handling
boto3==1.42.54
pillow==12.1.1
PyMuPDF==1.26.5

# Component library + forms
django-cotton>=1.0.0
django-template-partials>=24.4
django-tailwind>=3.8.0
django-crispy-forms>=2.3

# Static files
whitenoise==6.11.0
